Update Score task 5 Scores

+++ salvo/src/main/java/com/codeoftheweb/salvo/repository/GameRepository.java
@@ -8,6 +8,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface GameRepository extends JpaRepository<Game, Long> {
+public interface GameRepository extends JpaRepository<Game, Integer> {
     List<Game> findByCreationDate(Date creationDate);
 }
+
+
